backend/requirements.txt
fastapi
sqlmodel
uvicorn
pydantic
pydantic-settings
pytest
httpx
psycopg2-binary
pre-commit
python-jose[cryptography]
argon2-cffi
